% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloauto.R
\docType{data}
\name{alloauto}
\alias{alloauto}
\title{Leukemia data set}
\format{
This data frame contains the following columns:
  \itemize{
    \item time: Time to death or relapse, months
    \item type :Type of transplant (1=allogeneic, 2=autologous)
   \item delta:Leukemia-free survival indicator (0=alive without relapse, 1=dead
                                       or relapse)
 }
}
\source{
{
Klein and Moeschberger (1997) \emph{Survival Analysis Techniques for Censored
  and truncated data}, Springer.
 Kardaun Stat. Nederlandica 37 (1983), 103-126.
}
}
\description{
The alloauto data frame has 101 rows and 3 columns.
}
\examples{
{
data(alloauto)
str(alloauto)
}
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa,Christophe Chesneau, \email{abdisalam.hassan@amoud.edu.so}
}
\keyword{datasets}
