% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gastric.R
\docType{data}
\name{gastric}
\alias{gastric}
\title{Gastric data set}
\format{
This data frame contains the following columns:
  \itemize{
    \item time: survival times in days
    \item trt :treatments (1=chemotherapy + radiation; 0=chemotherapy alone)
   \item status:failure indicator (1=failure, 0=otherwise)
 }
}
\source{
{
Gastrointestinal Tumor Study Group. (1982) A Comparison of Combination Chemotherapy and Combined Modality Therapy for Locally Advanced Gastric Carcinoma. Cancer 49:1771-7.
}
}
\description{
The gastric data frame has 90 rows and variables.It is a data set from a clinical trial conducted by the Gastrointestinal Tumor Study Group (GTSG) in 1982. The data set refers to the survival times of patients with locally nonresectable gastric cancer. Patients were either treated with chemotherapy combined with radiation or chemotherapy alone.
}
\examples{
{
data(gastric)
str(gastric);head(gastric)
}
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa,Christophe Chesneau, \email{abdisalam.hassan@amoud.edu.so}
}
\keyword{datasets}
