% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larynx.R
\docType{data}
\name{larynx}
\alias{larynx}
\title{Larynx Cancer-Patients data set}
\format{
The data frame contains 90 rows and 5 columns:
 \itemize{
   \item time: time to event, in months
   \item delta: Censor indicator, 0 alive and 1 for dead
  \item stage: Stage of disease (1=stage 1, 2=stage2, 3=stage 3, 4=stage 4)
  \item diagyr: Year of diagnosis of larynx cancer
  \item age: Age at diagnosis of larynx cancer
}
}
\description{
Larynx Cancer-Patients data set which is widely used in the survival regression models
}
\references{
Klein and Moeschberger (1997) \emph{Survival Analysis Techniques for Censored
and truncated data}, Springer.
Kardaun Stat. Nederlandica 37 (1983), 103-126.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau, \email{abdisalam.hassan@amoud.edu.so}
}
\keyword{datasets}
