% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{pATLL}
\alias{pATLL}
\title{Arctangent-Log-logistic (ATLL)  Cumulative Distribution Function.}
\usage{
pATLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the ATLL Cumulative Distribution function
}
\description{
Arctangent-Log-logistic (ATLL)  Cumulative Distribution Function.
}
\examples{
t=runif(10,min=0,max=1)
pATLL(t=t, alpha=0.7, beta=0.5)

}
\references{
Alkhairy, I., Nagy, M., Muse, A. H., & Hussam, E. (2021). The Arctan-X family of distributions: Properties, simulation, and applications to actuarial sciences. Complexity, 2021.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau \email{abdisalam.hassan@amoud.edu.so}
}
