% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{pLN}
\alias{pLN}
\title{Lognormal (LN) Cumulative Distribution Function.}
\usage{
pLN(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: meanlog parameter}

\item{alpha}{: sdlog parameter}
}
\value{
the value of the LN cumulative distribution function
}
\description{
Lognormal (LN) Cumulative Distribution Function.
}
\examples{
t=runif(10,min=0,max=1)
pLN(t=t, kappa=0.75, alpha=0.95)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau   \email{abdisalam.hassan@amoud.edu.so}
}
