% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{pSCLL}
\alias{pSCLL}
\title{Secant-log-logistic (SCLL) Cumulative Distribution Function.}
\usage{
pSCLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the SCLL Cumulative Distribution function
}
\description{
Secant-log-logistic (SCLL) Cumulative Distribution Function.
}
\examples{
t=runif(10,min=0,max=1)
pSCLL(t=t, alpha=0.7, beta=0.5)

}
\references{
Souza, L., de Oliveira, W. R., de Brito, C. C. R., Chesneau, C., Fernandes, R., & Ferreira, T. A. (2022). Sec-G class of distributions: Properties and applications. Symmetry, 14(2), 299.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
