% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{pSLL}
\alias{pSLL}
\title{Sine-Log-logistic (SLL)  Cumulative Distribution Function.}
\usage{
pSLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the SLL Cumulative Distribution function
}
\description{
Sine-Log-logistic (SLL)  Cumulative Distribution Function.
}
\examples{
t=runif(10,min=0,max=1)
pSLL(t=t, alpha=0.7, beta=0.5)

}
\references{
Souza, L., Junior, W., De Brito, C., Chesneau, C., Ferreira, T., & Soares, L. (2019). On the Sin-G class of distributions: theory, model and application. Journal of Mathematical Modeling, 7(3), 357-379.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
