% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{pTLL}
\alias{pTLL}
\title{Tangent-Log-logistic (TLL)  Cumulative Distribution Function.}
\usage{
pTLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the TLL Cumulative Distribution function
}
\description{
Tangent-Log-logistic (TLL)  Cumulative Distribution Function.
}
\examples{
t=runif(10,min=0,max=1)
pTLL(t=t, alpha=0.7, beta=0.5)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau \email{abdisalam.hassan@amoud.edu.so}
}
