% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rASLL}
\alias{rASLL}
\title{Arcsine-Log-logistic (ASLL)  Hazard Rate Function.}
\usage{
rASLL(t, alpha, beta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the ASLL Hazard Rate  Function.
}
\description{
Arcsine-Log-logistic (ASLL)  Hazard Rate Function.
}
\examples{
t=runif(10,min=0,max=1)
rSLL(t=t, alpha=0.7, beta=0.5,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau \email{abdisalam.hassan@amoud.edu.so}
}
