% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rATLL}
\alias{rATLL}
\title{Arctangent-Log-logistic (ATLL)  Hazard  Function.}
\usage{
rATLL(t, alpha, beta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the ATLL hazard function
}
\description{
Arctangent-Log-logistic (ATLL)  Hazard  Function.
}
\examples{
t=runif(10,min=0,max=1)
rATLL(t=t, alpha=0.7, beta=0.5,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
