% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rCLL}
\alias{rCLL}
\title{Cosine-Log-logistic (CLL)  Hazard  Function.}
\usage{
rCLL(t, alpha, beta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the CLL hazard function
}
\description{
Cosine-Log-logistic (CLL)  Hazard  Function.
}
\examples{
t=runif(10,min=0,max=1)
rCLL(t=t, alpha=0.7, beta=0.5,log=FALSE)

}
\references{
Souza, L., Junior, W. R. D. O., de Brito, C. C. R., Ferreira, T. A., & Soares, L. G. (2019). General properties for the Cos-G class of distributions with applications. Eurasian Bulletin of Mathematics (ISSN: 2687-5632), 63-79.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
