% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rG}
\alias{rG}
\title{Gamma (G) Hazard Function.}
\usage{
rG(t, shape, scale, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the G hazard function
}
\description{
Gamma (G) Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
rG(t=t, shape=0.5, scale=0.85,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau   \email{abdisalam.hassan@amoud.edu.so}
}
