% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rMLL}
\alias{rMLL}
\title{Modified Log-logistic (MLL) hazard function.}
\usage{
rMLL(t, kappa, alpha, eta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the MLL hazard function
}
\description{
Modified Log-logistic (MLL) hazard function.
}
\examples{
t=runif(10,min=0,max=1)
rMLL(t=t, kappa=0.75, alpha=0.5, eta=0.9,log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
