% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rNGLL}
\alias{rNGLL}
\title{New Generalized Log-logistic (NGLL) hazard function.}
\usage{
rNGLL(t, kappa, alpha, eta, zeta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}

\item{zeta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the NGLL hazard function
}
\description{
New Generalized Log-logistic (NGLL) hazard function.
}
\examples{
t=runif(10,min=0,max=1)
rNGLL(t=t, kappa=0.5, alpha=0.35, eta=0.7, zeta=1.4, log=FALSE)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
