% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rSCLL}
\alias{rSCLL}
\title{Secant-log-logistic (SCLL) Hazard  Function.}
\usage{
rSCLL(t, alpha, beta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the SCLL hazard function
}
\description{
Secant-log-logistic (SCLL) Hazard  Function.
}
\examples{
t=runif(10,min=0,max=1)
rSCLL(t=t, alpha=0.7, beta=0.5,log=FALSE)

}
\references{
Souza, L., de Oliveira, W. R., de Brito, C. C. R., Chesneau, C., Fernandes, R., & Ferreira, T. A. (2022). Sec-G class of distributions: Properties and applications. Symmetry, 14(2), 299.

Tung, Y. L., Ahmad, Z., & Mahmoudi, E. (2021). The Arcsine-X Family of Distributions with Applications to Financial Sciences. Comput. Syst. Sci. Eng., 39(3), 351-363.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
