% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rSLL}
\alias{rSLL}
\title{Sine-Log-logistic (SLL)  Hazard  Function.}
\usage{
rSLL(t, alpha, beta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the SLL hazard function
}
\description{
Sine-Log-logistic (SLL)  Hazard  Function.
}
\examples{
t=runif(10,min=0,max=1)
rSLL(t=t, alpha=0.7, beta=0.5,log=FALSE)

}
\references{
Souza, L. (2015). New trigonometric classes of probabilistic distributions. esis, Universidade Federal Rural de Pernambuco, Brazil.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
