% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{rTLL}
\alias{rTLL}
\title{Tangent-Log-logistic (TLL)  Hazard  Function.}
\usage{
rTLL(t, alpha, beta, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}

\item{log}{:log scale (TRUE or FALSE)}
}
\value{
the value of the TLL hazard function
}
\description{
Tangent-Log-logistic (TLL)  Hazard  Function.
}
\examples{
t=runif(10,min=0,max=1)
rTLL(t=t, alpha=0.7, beta=0.5,log=FALSE)

}
\references{
Muse, A. H., Tolba, A. H., Fayad, E., Abu Ali, O. A., Nagy, M., & Yusuf, M. (2021). Modelling the COVID-19 mortality rate with a new versatile modification of the log-logistic distribution. Computational Intelligence and Neuroscience, 2021.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau \email{abdisalam.hassan@amoud.edu.so}
}
