% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sASLL}
\alias{sASLL}
\title{Arcsine-Log-logistic (ASLL)  Survival Function.}
\usage{
sASLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the ASLL Survival Function.
}
\description{
Arcsine-Log-logistic (ASLL)  Survival Function.
}
\examples{
t=runif(10,min=0,max=1)
sASLL(t=t, alpha=0.7, beta=0.5)

}
\references{
Tung, Y. L., Ahmad, Z., & Mahmoudi, E. (2021). The Arcsine-X Family of Distributions with Applications to Financial Sciences. Comput. Syst. Sci. Eng., 39(3), 351-363.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
