% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sCLL}
\alias{sCLL}
\title{Cosine-Log-logistic (CLL)  Survivor Function.}
\usage{
sCLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the CLL Survivor function
}
\description{
Cosine-Log-logistic (CLL)  Survivor Function.
}
\examples{
t=runif(10,min=0,max=1)
sCLL(t=t, alpha=0.7, beta=0.5)
}
\references{
Mahmood, Z., M Jawa, T., Sayed-Ahmed, N., Khalil, E. M., Muse, A. H., & Tolba, A. H. (2022). An Extended Cosine Generalized Family of Distributions for Reliability Modeling: Characteristics and Applications with Simulation Study. Mathematical Problems in Engineering, 2022.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau \email{abdisalam.hassan@amoud.edu.so}
}
