% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sG}
\alias{sG}
\title{Gamma (G) Survivor Function.}
\usage{
sG(t, shape, scale)
}
\arguments{
\item{t}{: positive argument}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}
}
\value{
the value of the G Survivor function
}
\description{
Gamma (G) Survivor Function.
}
\examples{
t=runif(10,min=0,max=1)
sG(t=t, shape=0.85, scale=0.5)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau   \email{abdisalam.hassan@amoud.edu.so}
}
