% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sLL}
\alias{sLL}
\title{Log-logistic  (LL) Survivor Function.}
\usage{
sLL(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}
}
\value{
the value of the LL survivor function
}
\description{
Log-logistic  (LL) Survivor Function.
}
\examples{
t=runif(10,min=0,max=1)
sLL(t=t, kappa=0.5, alpha=0.35)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
