% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sLN}
\alias{sLN}
\title{Lognormal (LN) Survivor Hazard Function.}
\usage{
sLN(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: meanlog parameter}

\item{alpha}{: sdlog parameter}
}
\value{
the value of the LN Survivor function
}
\description{
Lognormal (LN) Survivor Hazard Function.
}
\examples{
t=runif(10,min=0,max=1)
sLN(t=t, kappa=0.75, alpha=0.95)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau   \email{abdisalam.hassan@amoud.edu.so}
}
