% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sMKW}
\alias{sMKW}
\title{Modified Kumaraswamy Weibull (MKW) Survivor Function.}
\usage{
sMKW(t, alpha, kappa, eta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: Inverse scale parameter}

\item{kappa}{: shape parameter}

\item{eta}{: shape parameter}
}
\value{
the value of the MKW survivor function
}
\description{
Modified Kumaraswamy Weibull (MKW) Survivor Function.
}
\examples{
t=runif(10,min=0,max=1)
sMKW(t=t,alpha=0.35, kappa=0.7, eta=1.4)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
