% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sMLL}
\alias{sMLL}
\title{Modified Log-logistic (MLL) survivor function.}
\usage{
sMLL(t, kappa, alpha, eta)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}

\item{eta}{: shape parameter}
}
\value{
the value of the MLL survivor function
}
\description{
Modified Log-logistic (MLL) survivor function.
}
\examples{
t=runif(10,min=0,max=1)
sMLL(t=t, kappa=0.75, alpha=0.5, eta=0.9)

}
\references{
Kayid, M. (2022). Applications of Bladder Cancer Data Using a Modified Log-Logistic Model. Applied Bionics and Biomechanics, 2022.
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau  \email{abdisalam.hassan@amoud.edu.so}
}
