% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sSCLL}
\alias{sSCLL}
\title{Secant-log-logistic (SCLL) Survivor Function.}
\usage{
sSCLL(t, alpha, beta)
}
\arguments{
\item{t}{: positive argument}

\item{alpha}{: scale parameter}

\item{beta}{: shape parameter}
}
\value{
the value of the SCLL Survivor function
}
\description{
Secant-log-logistic (SCLL) Survivor Function.
}
\examples{
t=runif(10,min=0,max=1)
sSCLL(t=t, alpha=0.7, beta=0.5)
}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau \email{abdisalam.hassan@amoud.edu.so}
}
