% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{sW}
\alias{sW}
\title{Weibull  (W) Survivor Function.}
\usage{
sW(t, kappa, alpha)
}
\arguments{
\item{t}{: positive argument}

\item{kappa}{: scale parameter}

\item{alpha}{: shape parameter}
}
\value{
the value of the W Survivor function
}
\description{
Weibull  (W) Survivor Function.
}
\examples{
t=runif(10,min=0,max=1)
sW(t=t, kappa=0.75, alpha=0.5)

}
\author{
Abdisalam Hassan Muse, Samuel Mwalili, Oscar Ngesa, Christophe Chesneau   \email{abdisalam.hassan@amoud.edu.so}
}
